Upgrade to Swift 3 and factor out API keys
The app currently compiles after migrating to Swift 3 and the API keys
have been moved out of the constants file to be imported from the
Keys.xcconfig file.

There's probably a bunch of really awful code generated by the migration
(and also written in the past) that will need to be refactored. There's
also probably broken functionality that has to be tested.

Also added a linter and ran autocorrect once. Rules with a lot of
violations are temporarily disabled and will have to be re-enabled as
more files are refactored.